Camera related stuff moved to fragment. Again. Things seem fine.
