+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
          android:name="android.support.FILE_PROVIDER_PATHS"
          android:resource="@xml/file_provider_paths"/>
          </provider>
-        <activity android:name=".activities.CameraActivity">
+        <activity android:name=".camera.CameraActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
