+++ app/src/main/res/layout/activity_camera.xml
@@ -6,51 +6,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
-
-    <com.illegaldisease.banreco.camera.CameraSourcePreview
-        android:id="@+id/preview"
+    <fragment
+        android:name="com.illegaldisease.banreco.camera.CameraFragment"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" >
+        android:layout_height="match_parent"
+        android:id="@+id/fragmentCamera" />
 
-        <com.illegaldisease.banreco.camera.GraphicOverlay
-            android:id="@+id/graphicOverlay"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    </com.illegaldisease.banreco.camera.CameraSourcePreview>
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab_take_photo"
-        style="@style/fab"
-        android:layout_width="56dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        app:srcCompat="@drawable/ic_camera"
-        android:visibility="invisible"
-        app:fabSize="normal" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab_flash"
-        style="@style/fab"
-        android:layout_width="56dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginEnd="18dp"
-        app:srcCompat="@drawable/ic_flash_on_white_24px"
-        android:visibility="invisible"
-        app:fabSize="normal" />
-
-    <ProgressBar
-        android:id="@+id/indeterminateBar"
-        style="?android:attr/progressBarStyleSmall"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="center"
-        android:layout_marginTop="215dp"
-        android:indeterminate="true" />
 </RelativeLayout>
